Extract ObservationDetailView header to a new component
