+++ web/src/components/trace2/components/ObservationDetailView/ObservationDetailView.tsx
@@ -17,14 +17,7 @@
  * - View mode toggle changes
  */
 
-import {
-  type ObservationType,
-  AnnotationQueueObjectType,
-  isGenerationLike,
-} from "@langfuse/shared";
 import { type ObservationReturnTypeWithMetadata } from "@/src/server/api/routers/traces";
-import { ItemBadge } from "@/src/components/ItemBadge";
-import { LocalIsoDate } from "@/src/components/LocalIsoDate";
 import {
   TabsBar,
   TabsBarContent,
@@ -33,17 +26,6 @@ import {
 } from "@/src/components/ui/tabs-bar";
 import { Tabs, TabsList, TabsTrigger } from "@/src/components/ui/tabs";
 import { useMemo, useState } from "react";
-import {
-  LatencyBadge,
-  TimeToFirstTokenBadge,
-  EnvironmentBadge,
-  VersionBadge,
-  LevelBadge,
-  StatusMessageBadge,
-} from "./ObservationMetadataBadgesSimple";
-import { CostBadge, UsageBadge } from "./ObservationMetadataBadgesTooltip";
-import { ModelBadge } from "./ObservationMetadataBadgeModel";
-import { ModelParametersBadges } from "./ObservationMetadataBadgeModelParameters";
 import ScoresTable from "@/src/components/table/use-cases/scores";
 import { IOPreview } from "@/src/components/trace2/components/IOPreview/IOPreview";
 import { useJsonExpansion } from "@/src/components/trace2/contexts/JsonExpansionContext";
@@ -51,16 +33,12 @@ import { useMedia } from "@/src/components/trace2/api/useMedia";
 import { useSelection } from "@/src/components/trace2/contexts/SelectionContext";
 import { useViewPreferences } from "@/src/components/trace2/contexts/ViewPreferencesContext";
 
-// Header action components
-import { CopyIdsPopover } from "@/src/components/trace2/components/_shared/CopyIdsPopover";
-import { NewDatasetItemFromExistingObject } from "@/src/features/datasets/components/NewDatasetItemFromExistingObject";
-import { AnnotateDrawer } from "@/src/features/scores/components/AnnotateDrawer";
-import { CreateNewAnnotationQueueItem } from "@/src/features/annotation-queues/components/CreateNewAnnotationQueueItem";
-import { CommentDrawerButton } from "@/src/features/comments/CommentDrawerButton";
-import { JumpToPlaygroundButton } from "@/src/features/playground/page/components/JumpToPlaygroundButton";
-import { PromptBadge } from "@/src/components/trace2/components/_shared/PromptBadge";
+// Contexts and hooks
 import { useTraceData } from "@/src/components/trace2/contexts/TraceDataContext";
 import { useParsedObservation } from "@/src/hooks/useParsedObservation";
+
+// Extracted components
+import { ObservationDetailViewHeader } from "./ObservationDetailViewHeader";
 
 export interface ObservationDetailViewProps {
   observation: ObservationReturnTypeWithMetadata;
@@ -147,137 +125,18 @@ export function ObservationDetailView({
     return null;
   }, [observation.latency, observation.startTime, observation.endTime]);
 
-  // Format cost and usage values
-  const totalCost = observation.totalCost;
-  const totalUsage = observation.totalUsage;
-  const inputUsage = observation.inputUsage;
-  const outputUsage = observation.outputUsage;
-
   return (
     <div className="flex h-full flex-col overflow-hidden">
-      {/* Header section */}
-      <div className="flex-shrink-0 space-y-2 border-b p-2">
-        {/* Title row with actions */}
-        <div className="grid w-full grid-cols-1 items-start gap-2 @2xl:grid-cols-[auto,auto] @2xl:justify-between">
-          <div className="flex w-full flex-row items-start gap-1">
-            <div className="mt-1.5">
-              <ItemBadge type={observation.type as ObservationType} isSmall />
-            </div>
-            <span className="mb-0 ml-1 line-clamp-2 min-w-0 break-all font-medium md:break-normal md:break-words">
-              {observation.name || observation.id}
-            </span>
-            <CopyIdsPopover
-              idItems={[
-                { id: traceId, name: "Trace ID" },
-                { id: observation.id, name: "Observation ID" },
-              ]}
-            />
-          </div>
-          {/* Action buttons */}
-          <div className="flex h-full flex-wrap content-start items-start justify-start gap-0.5 @2xl:mr-1 @2xl:justify-end">
-            {observationWithIOCompat.data && (
-              <NewDatasetItemFromExistingObject
-                traceId={traceId}
-                observationId={observation.id}
-                projectId={projectId}
-                input={observationWithIOCompat.data.input}
-                output={observationWithIOCompat.data.output}
-                metadata={observationWithIOCompat.data.metadata}
-                key={observation.id}
-                size="sm"
-              />
-            )}
-            <div className="flex items-start">
-              <AnnotateDrawer
-                key={"annotation-drawer-" + observation.id}
-                projectId={projectId}
-                scoreTarget={{
-                  type: "trace",
-                  traceId: traceId,
-                  observationId: observation.id,
-                }}
-                scores={observationScores}
-                scoreMetadata={{
-                  projectId: projectId,
-                  environment: observation.environment,
-                }}
-                size="sm"
-              />
-              <CreateNewAnnotationQueueItem
-                projectId={projectId}
-                objectId={observation.id}
-                objectType={AnnotationQueueObjectType.OBSERVATION}
-                size="sm"
-              />
-            </div>
-            {observationWithIOCompat.data &&
-              isGenerationLike(observationWithIOCompat.data.type) && (
-                <JumpToPlaygroundButton
-                  source="generation"
-                  generation={observationWithIOCompat.data}
-                  analyticsEventName="trace_detail:test_in_playground_button_click"
-                  size="sm"
-                />
-              )}
-            <CommentDrawerButton
-              projectId={projectId}
-              objectId={observation.id}
-              objectType="OBSERVATION"
-              count={comments.get(observation.id)}
-              size="sm"
-            />
-          </div>
-        </div>
-
-        {/* Metadata badges */}
-        <div className="flex flex-col gap-1">
-          {/* Timestamp on its own row */}
-          <div className="flex items-center">
-            <LocalIsoDate
-              date={observation.startTime}
-              accuracy="millisecond"
-              className="text-xs"
-            />
-          </div>
-          {/* Other badges on second row */}
-          <div className="flex flex-wrap items-center gap-1">
-            <LatencyBadge latencySeconds={latencySeconds} />
-            <TimeToFirstTokenBadge
-              timeToFirstToken={observation.timeToFirstToken}
-            />
-            <EnvironmentBadge environment={observation.environment} />
-            <CostBadge
-              totalCost={totalCost}
-              costDetails={observation.costDetails}
-            />
-            <UsageBadge
-              type={observation.type}
-              inputUsage={inputUsage}
-              outputUsage={outputUsage}
-              totalUsage={totalUsage}
-              usageDetails={observation.usageDetails}
-            />
-            <VersionBadge version={observation.version} />
-            <ModelBadge
-              model={observation.model}
-              internalModelId={observation.internalModelId}
-              projectId={projectId}
-              usageDetails={observation.usageDetails}
-            />
-            <ModelParametersBadges
-              modelParameters={observation.modelParameters}
-            />
-            <LevelBadge level={observation.level} />
-            <StatusMessageBadge statusMessage={observation.statusMessage} />
-            {observation.promptId && (
-              <PromptBadge
-                promptId={observation.promptId}
-                projectId={projectId}
-              />
-            )}
-          </div>
-        </div>
-      </div>
+      {/* Header section (extracted component) */}
+      <ObservationDetailViewHeader
+        observation={observation}
+        observationWithIO={observationWithIO}
+        projectId={projectId}
+        traceId={traceId}
+        latencySeconds={latencySeconds}
+        observationScores={observationScores}
+        commentCount={comments.get(observation.id)}
+      />
 
       {/* Tabs section */}
       <TabsBar
